requirements.txt
genkit-plugin-google-genai
pillow